added logging, changed type of size field to size_t, changed buffer name target to result_shape in metal kernels

+++ include/memory_pool.h
@@ -9,11 +9,11 @@ class MemoryPool {
 private:
   std::multiset<std::shared_ptr<Memory>> available_pool;
   std::multiset<std::shared_ptr<Memory>> used_pool;
-  int _compute_pool_size(int requested_size);
+  size_t _compute_pool_size(size_t requested_size);
 
 public:
-  std::shared_ptr<Memory> request_memory(DeviceType device, int size,
+  std::shared_ptr<Memory> request_memory(DeviceType device, size_t size,
                                          DType dtype);
-  std::shared_ptr<Memory> find_suitable_block(int requested);
+  std::shared_ptr<Memory> find_suitable_block(size_t requested);
   void return_memory(std::shared_ptr<Memory> memory);
 };
